requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# pil_video/pil_video.py: 8
Pillow == 8.0.1

# pil_video/pil_video.py: 11
imageio == 2.9.0
